refactor(mobile): moved discard logic to include title
