+++ mobile/src/ui/Modal/ui/screens/Modal/ModalWrapper.tsx
@@ -7,14 +7,18 @@ import {SafeAreaViewProps} from 'react-native-safe-area-context'
 import {SafeArea} from '~/ui/SafeArea/SafeArea'
 import {Text} from '~/ui/Text/Text'
 
+import {Discard} from '../../shared/Discard'
+
 type Props = SafeAreaViewProps & {
   title?: string
   footer?: React.ReactNode
+  onDragYChange: (y: number) => void
 }
 
 export const ModalWrapper = ({
   title,
   footer,
+  onDragYChange,
   children,
   edges,
   ...rest
@@ -23,13 +27,15 @@ export const ModalWrapper = ({
 
   return (
     <SafeArea edges={edges} style={[a.justify_between]} {...rest}>
-      {title && (
-        <View style={[a.py_sm, a.px_lg]}>
-          <Text style={[a.heading_3_medium, ta.text_gray_max, a.text_center]}>
-            {title}
-          </Text>
-        </View>
-      )}
+      <Discard.Indicator onDragYChange={onDragYChange}>
+        {title && (
+          <View style={[a.py_sm, a.px_lg]}>
+            <Text style={[a.heading_3_medium, ta.text_gray_max, a.text_center]}>
+              {title}
+            </Text>
+          </View>
+        )}
+      </Discard.Indicator>
 
       {children}
